Makes log creation safer with nil Rails loggers.
On a pure Rack app that uses some Rails based middleware like Devise, Rails.logger
can be defined, but nil. In that case, the safest behavior is probably to continue
on until we return a new Logger to STDERR.

This is especially true on Heroku where we will definitely end up in the rescue
block.